% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipio_comparativa_descriptivos.R
\name{municipio_comparativa_descriptivos}
\alias{municipio_comparativa_descriptivos}
\title{Integra como telemetría al dispositivo Comparativa - Indicadores descriptivos los datos desciptivos (BBDD MITMA) del municipio demandado}
\usage{
municipio_comparativa_descriptivos(
  ip_plataforma,
  id_dispositivo_comparativa,
  nombre_municipio
)
}
\arguments{
\item{ip_plataforma, }{id_dispositivo_comparativa, nombre_municipio}
}
\value{
json
}
\description{
Integra como telemetría al dispositivo Comparativa - Indicadores descriptivos los datos desciptivos (BBDD MITMA) del municipio demandado
}
\examples{
 municipio_comparativa_descriptivos("http://82.223.66.83:8080","74041e60-fc46-11ec-94d8-c1cee3606f2e","Barakaldo")

}
