% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipios_espana.R
\name{municipios_espana}
\alias{municipios_espana}
\title{Devuelve un json con los nombres de los municipios de España}
\usage{
municipios_espana(valor = "1")
}
\arguments{
\item{valor}{}
}
\value{
json
}
\description{
Devuelve un json con los nombres de los municipios de España
}
\examples{
 municipios_espana("1")

}
